Make all stats saved in doomretro.cfg 64-bit
You can now kill up to 18,446,744,073,709,551,615 cacodemons!

+++ src/d_main.c
@@ -166,7 +166,7 @@ dboolean            regenhealth;
 dboolean            respawnitems;
 dboolean            respawnmonsters;        // checkparm of -respawn
 
-unsigned int        stat_runs = 0;
+uint64_t            stat_runs = 0;
 
 skill_t             startskill;
 int                 startepisode;
